backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4   
bcrypt==4.0.1            
